Fix regex and stuff
