+++ test/.babelrc
@@ -2,7 +2,6 @@
   "presets": [
     "flow",
     "env",
-    "stage-0",
     "react",
     ["./babel-preset-emotion-test", { "inline": true }]
   ]
